modify something and add something
